[U]更新 `会员` uid从10000起 防止跟管理员冲突

+++ app/patch/files/db.2018041717.php
@@ -3,12 +3,20 @@
 defined('iPHP') OR require (dirname(__FILE__).'/../../../iCMS.php');
 
 return patch::upgrade(function(){
-  iDB::query("
-  ALTER TABLE `#iCMS@__user`
-    DROP INDEX `username`,
-    DROP INDEX `email`,
-    ADD  KEY `username` (`username`);
-  ");
+  $indexs = apps_db::indexes('#iCMS@__user');
+  if($indexs['email']){
+    iDB::query("
+    ALTER TABLE `#iCMS@__user`
+      DROP INDEX `email`;
+    ");
+  }
+  if($indexs['username']){
+    iDB::query("
+    ALTER TABLE `#iCMS@__user`
+      DROP INDEX `username`,
+      ADD  KEY `username` (`username`);
+    ");
+  }
   return '更新用户表索引';
 });
